Changed app file structure, separated components out into modules.
Changed app file structure, separated components out into modules.
