+++ src/api/api.ts
@@ -1,5 +1,5 @@
 import { OperationVariables, QueryOptions, ApolloQueryResult } from 'apollo-boost';
-import { apolloClient } from '../containers/App/App';
+import { apolloClient } from '../modules/core/containers/App/App';
 
 const query = <T = any, TVariables = OperationVariables>(options: QueryOptions<TVariables>): Promise<ApolloQueryResult<T>> => {
   return apolloClient.query<T, TVariables>(options);
