+++ src/modules/core/components/NavigationItem/NavigationItem.scss
