+++ src/modules/core/containers/App/App.scss
