+++ src/modules/core/containers/App/App.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import { Provider } from 'react-redux';
-import store from '../../redux/configureStore';
+import store from '../../../../redux/configureStore';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
