+++ src/modules/seeker/components/JobDetails/JobDetails.scss
