+++ src/modules/seeker/containers/MatchedJobsPage/MatchedJobsPage.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
-import pageWrapper from '../../components/PageWrapper/PageWrapper';
+import pageWrapper from '../../../shared/components/PageWrapper/PageWrapper';
 import './MatchedJobsPage.scss';
-import Alert from '../../components/Alert/Alert';
-import noJobsImg from '../../images/empty.svg';
-import Loading from '../../components/Loading/Loading';
-import Error from '../../components/Error/Error';
+import Alert from '../../../shared/components/Alert/Alert';
+import noJobsImg from '../../../../images/empty.svg';
+import Loading from '../../../shared/components/Loading/Loading';
+import Error from '../../../shared/components/Error/Error';
 import { useQuery } from '@apollo/react-hooks';
 import { gql } from 'apollo-boost';
 import { useSelector } from 'react-redux';
-import { IAppState } from '../../redux/appState';
-import { IUser } from '../../models/User';
+import { IAppState } from '../../../../redux/appState';
+import { IUser } from '../../../../models/User';
 
 const MatchedJobsPage = () => {
   const user: IUser | null = useSelector((state: IAppState) => state.authentication.user);
