+++ src/modules/seeker/containers/PotentialJobDetailsPage/PotentialJobDetailsPage.scss
