+++ src/modules/shared/components/Card/Card.scss
