+++ src/modules/shared/components/Error/Error.scss
