+++ src/modules/shared/components/Loading/Loading.scss
