+++ src/modules/shared/components/Loading/Loading.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import loading from '../../images/loading.svg';
+import loading from '../../../../images/loading.svg';
 import './Loading.scss';
 
 const Loading = () => {
