+++ src/modules/shared/components/Pill/Pill.scss
